#13 Config multer, cloudinary and upload image product

+++ server/controllers/coupon.js
@@ -3,7 +3,12 @@ const asyncHandler = require('express-async-handler');
 
 var that = (module.exports = {
   createCoupon: asyncHandler(async (req, res) => {
-    const result = await Coupon.create(req.body);
+    const { name, discount, expiry } = req.body;
+    const ONE_DAY = 24 * 60 * 60 * 1000;
+    const result = await Coupon.create({
+      ...req.body,
+      expiry: Date.now() + expiry * ONE_DAY,
+    });
     return res.json({
       success: result ? true : false,
       createdCoupon: result ? result : 'Cannot create Coupon',
@@ -11,7 +16,7 @@ var that = (module.exports = {
   }),
 
   getAllCoupons: asyncHandler(async (req, res) => {
-    const result = await Coupon.find();
+    const result = await Coupon.find().select('-createdAt -updatedAt');
     return res.json({
       success: result ? true : false,
       allCoupons: result ? result : 'Cannot create Coupon',
@@ -19,7 +24,11 @@ var that = (module.exports = {
   }),
 
   updateCoupon: asyncHandler(async (req, res) => {
+    const { name, discount, expiry } = req.body;
     const { cid } = req.params;
+    const ONE_DAY = 24 * 60 * 60 * 1000;
+    if (Object.keys(cid).length === 0) throw new Error('Missing input!');
+    if (expiry) expiry = Date.now() + ONE_DAY * expiry;
     const result = await Coupon.findByIdAndUpdate(cid, req.body, {
       new: true,
     });
@@ -31,6 +40,7 @@ var that = (module.exports = {
 
   deleteCoupon: asyncHandler(async (req, res) => {
     const { cid } = req.params;
+    if (Object.keys(cid).length === 0) throw new Error('Missing input!');
     const result = await Coupon.findByIdAndDelete(cid);
     return res.json({
       success: result ? true : false,
